move app.tsx inside src
